refactor: domain 클래스들을 model 패키지로 변경하라
