+++ src/main/java/racingcar/model/Car.java
@@ -1,6 +1,6 @@
-package racingcar.service;
+package racingcar.model;
 
-import static racingcar.config.RacingConstants.MAX_LENGTH_OF_CAR_NAME;
+import static racingcar.model.RacingConstants.MAX_LENGTH_OF_CAR_NAME;
 import static racingcar.View.ViewConstants.MOVE_SYMBOL;
 import static racingcar.View.ViewConstants.NAME_LENGTH_ERROR_MESSAGE;
 import static racingcar.View.ViewConstants.RACE_DISPLAY_FORMAT;
